api/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.0
aiosignal==1.3.2
alembic==1.16.1
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
async-timeout==4.0.3
asyncio==3.4.3
attrs==25.3.0
azure-ai-documentintelligence==1.0.2
azure-ai-formrecognizer==3.3.3
azure-common==1.1.28
azure-core==1.34.0
azure-core-tracing-opentelemetry==1.0.0b12
azure-cosmos==4.9.0
azure-identity==1.23.0
azure-monitor-opentelemetry==1.6.9
azure-monitor-opentelemetry-exporter==1.0.0b36
azure-monitor-query==1.4.1
azure-search-documents==11.5.2
azure-storage-blob==12.25.1
azure-keyvault-secrets==4.9.0
backoff==2.2.1
beautifulsoup4==4.13.4
bs4==0.0.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
cryptography==45.0.3
dataclasses-json==0.6.7
Deprecated==1.2.18
distro==1.9.0
dnspython==2.7.0
email_validator==2.2.0
et_xmlfile==2.0.0
exceptiongroup==1.2.2
fastapi==0.115.12
fastapi-cli==0.0.7
fixedint==0.1.6
frozenlist==1.6.0
glob2==0.7
greenlet==3.2.2
gunicorn==23.0.0
h11==0.14.0
html2text==2025.4.15
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
img2pdf==0.6.1
importlib_metadata==8.6.1
inquirerpy==0.3.4
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.25
langchain-community==0.3.24
langchain-core==0.3.61
langchain-openai==0.3.18
langchain-text-splitters==0.3.8
langchainhub==0.1.21
langsmith==0.3.42
lxml==5.4.0
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mdurl==0.1.2
msal==1.32.3
msal-extensions==1.3.1
msrest==0.7.1
multidict==6.4.4
mypy_extensions==1.1.0
numpy==2.0.2
oauthlib==3.2.2
openai==1.82.0
openpyxl==3.1.5
opentelemetry-api==1.31.1
opentelemetry-instrumentation==0.52b1
opentelemetry-instrumentation-asgi==0.52b1
opentelemetry-instrumentation-dbapi==0.52b1
opentelemetry-instrumentation-django==0.52b1
opentelemetry-instrumentation-fastapi==0.52b1
opentelemetry-instrumentation-flask==0.52b1
opentelemetry-instrumentation-logging==0.52b1
opentelemetry-instrumentation-psycopg2==0.52b1
opentelemetry-instrumentation-requests==0.52b1
opentelemetry-instrumentation-urllib==0.52b1
opentelemetry-instrumentation-urllib3==0.52b1
opentelemetry-instrumentation-wsgi==0.52b1
opentelemetry-resource-detector-azure==0.1.5
opentelemetry-sdk==1.31.1
opentelemetry-semantic-conventions==0.52b1
opentelemetry-util-http==0.52b1
orjson==3.10.15
packaging==24.2
passlib==1.7.4
pdfminer.six==20250506
pfzy==0.3.4
pikepdf==9.7.0
pillow==11.2.1
prompt_toolkit==3.0.51
prometheus-client==0.21.1
propcache==0.3.1
psutil==6.1.1
pycparser==2.22
pydantic==2.11.5
pydantic-extra-types==2.10.2
pydantic-settings==2.7.1
pydantic_core==2.33.2
Pygments==2.19.1
PyJWT==2.10.1
pymongo==4.13.0
PyMuPDF==1.26.0
pymupdf4llm==0.0.24
PyMySQL==1.1.1
pypdf==5.5.0
python-docx==1.1.2
python-dotenv==1.0.1
python-json-logger==3.3.0
python-multipart==0.0.20
python-pptx==1.0.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.4
rich-toolkit==0.13.2
scikit-learn==1.6.1
scipy==1.13.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
starlette==0.41.3
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.9.0
toml==0.10.2
tomli==2.2.1
tqdm==4.67.1
typer==0.15.1
types-requests==2.32.0.20250515
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.12.2
ujson==5.10.0
ulid-py==1.1.0
urllib3==2.4.0
uvicorn==0.32.0
uvloop==0.21.0
watchfiles==1.0.4
wcwidth==0.2.13
websockets==14.2
wrapt==1.17.2
XlsxWriter==3.2.3
yarl==1.20.0
zipp==3.22.0
zstandard==0.23.0
